# docs/flatmates-openapi.yaml
openapi: 3.1.0
info:
  title: 360 Flatmates API
  description: |
    Curated API contract for the 360 Flatmates surfaces (web + mobile) — a swipe-based
    flatmate-finding app for India's young professionals.

    This is a hand-maintained, flatmates-focused subset of the full backend OpenAPI spec.
    The authoritative full spec is `docs/openapi.json` in `360ghar-backend`, generated by
    `scripts/generate_openapi.py`. Every path here MUST exist in that spec — if a path in
    this file is not in `docs/openapi.json`, it is stale and must be removed. The mirror
    copy in `360-flatmates-web/docs/flatmates-openapi.yaml` is kept byte-identical to this
    file.

    ## Authentication
    All endpoints require authentication via Bearer token in the `Authorization` header, unless marked as public.
    Tokens are obtained via Supabase Auth (Phone OTP + Password).

    ## User Modes
    - **room_poster** — Has a flat, wants to rent out a spare room
    - **co_hunter** — Looking for people to flat-search with
    - **open_to_both** — Flexible, open to either

    ## Base URL
    `/api/v1`
  version: 1.1.0
  contact:
    name: 360 Ghar
    url: https://360ghar.com
  license:
    name: Confidential
    url: https://360ghar.com
servers:
  - url: https://api.360ghar.com/api/v1
    description: Production
  - url: http://localhost:8000/api/v1
    description: Local Development
...
paths:
...
  /properties:
    post:
      tags: [Listings]
      summary: Create a listing
      description: |
        Create a new property listing. For flatmate listings, set `property_type=flatmate` and `purpose=rent`.
        The listing enters manual review (24-hour SLA) before going live.
        AI pre-screening runs automatically before human review.
      operationId: createListing
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PropertyCreate'
      responses:
        '201':
          description: Listing created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Property'
        '400':
          $ref: '#/components/responses/BadRequest'

    get:
      tags: [Listings]
      summary: Discover listings
      description: |
        Discover listings with comprehensive filtering.
        For flatmate listings, filter with `property_type=flatmate` and `purpose=rent`.
        Supports location-based search, text search, semantic search, and 30+ filter parameters.
      operationId: discoverListings
      security: []
      parameters:
        - name: property_type
          in: query
          schema:
            type: array
            items:
              type: string
              enum: [flatmate, pg]
          description: Filter by property type
        - name: purpose
          in: query
          schema:
            type: string
            enum: [rent]
          description: Filter by purpose
        - name: lat
          in: query
          schema:
            type: number
            format: float
          description: Latitude for location-based search
        - name: lng
          in: query
          schema:
            type: number
            format: float
          description: Longitude for location-based search
        - name: radius
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 5
          description: Search radius in km
        - name: q
          in: query
          schema:
            type: string
          description: Search query for text or semantic search
        - name: semantic_search
          in: query
          schema:
            type: boolean
            default: false
          description: Enable semantic vector similarity search
        - name: city
          in: query
          schema:
            type: string
        - name: locality
          in: query
          schema:
            type: string
        - name: price_min
          in: query
          schema:
            type: number
            minimum: 0
        - name: price_max
          in: query
          schema:
            type: number
            maximum: 1000000000
        - name: bedrooms_min
          in: query
          schema:
            type: integer
            minimum: 0
        - name: bedrooms_max
          in: query
          schema:
            type: integer
            maximum: 20
        - name: sharing_type
          in: query
          schema:
            type: string
            enum: [private_room, shared_room, master_bedroom, entire_flat]
        - name: gender_preference
          in: query
          schema:
            type: string
            enum: [male, female, any]
        - name: move_in
          in: query
          schema:
            type: string
            enum: [immediate, this_month, next_month, flexible]
          description: Move-in timeline filter
        - name: available_from
          in: query
          schema:
            type: string
            format: date
          description: Minimum availability date (YYYY-MM-DD)
        - name: amenities
          in: query
          schema:
            type: array
            items:
              type: string
        - name: features
          in: query
          schema:
            type: array
            items:
              type: string
        - name: sort_by
          in: query
          schema:
            type: string
            enum: [distance, price_low, price_high, newest, popular, relevance]
            default: newest
        - name: exclude_swiped
          in: query
          schema:
            type: boolean
            default: false
          description: Exclude properties already swiped by the authenticated user
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Paginated listings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnifiedPropertyResponse'
  /properties/semantic-search:
    get:
      tags: [Listings]
      summary: Semantic property search
      description: Perform semantic (vector-powered) property search combining vector similarity with traditional filters.
      operationId: semanticSearchListings
      security: []
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
          description: Search query (required for semantic search)
        - name: property_type
          in: query
          schema:
            type: array
            items:
              type: string
              enum: [flatmate, pg]
        - name: city
          in: query
          schema:
            type: string
        - name: locality
          in: query
          schema:
            type: string
        - name: price_min
          in: query
          schema:
            type: number
            minimum: 0
        - name: price_max
          in: query
          schema:
            type: number
            maximum: 1000000000
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Paginated semantic search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnifiedPropertyResponse'
        '400':
          description: Search query required
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
                    example: A search query (q) is required for semantic search
  /properties/recommendations:
    get:
      tags: [Listings]
      summary: Get property recommendations
      description: |
        Get property recommendations.
        - With authentication: Personalized recommendations based on user preferences and swipes
        - Without authentication: Popular properties based on likes and recency
      operationId: getPropertyRecommendations
      security:
        - BearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 10
      responses:
        '200':
          description: Recommended properties
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Property'
  /properties/me:
    get:
      tags: [Listings]
      summary: Get my listings
      description: List all listings owned by the current user.
      operationId: getMyListings
      security:
        - BearerAuth: []
      responses:
        '200':
          description: User's listings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Property'
  /properties/{property_id}:
    get:
      tags: [Listings]
      summary: Get listing details
      description: |
        Get full property details. For flatmate listings, non-live listings are only visible to the owner or admins.
        Enriched with user-specific context (liked status, visit info) when authenticated.
      operationId: getListingDetails
      security: []
      parameters:
        - name: property_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Property details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Property'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Listings]
      summary: Update a listing
      description: |
        Update a listing. Owners can update their own listings.
        To pause/resume a listing, update `listing_preferences.moderation_status`.
      operationId: updateListing
      security:
        - BearerAuth: []
      parameters:
        - name: property_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PropertyUpdate'
      responses:
        '200':
          description: Updated listing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Property'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Listings]
      summary: Delete a listing
      operationId: deleteListing
      security:
        - BearerAuth: []
      parameters:
        - name: property_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Listing deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Property deleted successfully
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ===================== NOTIFICATIONS =====================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase Auth JWT token
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    PropertyCreate:
      type: object
      required: [property_type, purpose, title, city, locality, monthly_rent]
      properties:
        property_type:
          $ref: '#/components/schemas/PropertyType'
        purpose:
          $ref: '#/components/schemas/PropertyPurpose'
        title:
          type: string
          maxLength: 200
        description:
          type: string
          maxLength: 5000
        city:
          type: string
        state:
          type: string
        locality:
          type: string
        sub_locality:
          type: string
        address:
          type: string
          description: Full address (used for geocoding, not displayed publicly)
        latitude:
          type: number
          format: float
        longitude:
          type: number
          format: float
        monthly_rent:
          type: number
          minimum: 0
        security_deposit:
          type: number
          minimum: 0
        maintenance_charges:
          type: number
          minimum: 0
        area_sqft:
          type: number
          minimum: 0
        bedrooms:
          type: integer
          minimum: 0
        bathrooms:
          type: integer
          minimum: 0
        features:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
        main_image_url:
          type: string
          format: uri
        image_urls:
          type: array
          items:
            type: string
            format: uri
        available_from:
          type: string
          format: date
        gender_preference:
          type: string
          enum: [male, female, any]
        sharing_type:
          $ref: '#/components/schemas/ListingSharingType'
        video_tour_url:
          type: string
          format: uri
        society_type:
          type: string
          enum: [gated, independent, co_living, pg]
        society_amenities:
          type: array
          items:
            type: string
        society_vibe_tags:
          type: array
          items:
            type: string
        listing_preferences:
          type: object
          additionalProperties: true
          description: Listing-specific preferences and metadata
...
    UnifiedPropertyResponse:
      type: object
      properties:
        properties:
          type: array
          items:
            $ref: '#/components/schemas/Property'
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        total_pages:
          type: integer
        filters_applied:
          type: object
          additionalProperties: true
        search_center:
          type: object
          properties:
            latitude:
              type: number
            longitude:
              type: number
...
    PropertyType:
      type: string
      enum: [pg, flatmate]
    PropertyPurpose:
      type: string
      enum: [rent]
    ListingSharingType:
      type: string
      enum: [private_room, shared_room, master_bedroom, entire_flat]

    # ===================== ERROR SCHEMA =====================
    Error:
      type: object
      properties:
        detail:
          type: string
          description: Error message
        error_code:
          type: string
          description: Optional machine-readable error code